Former-commit-id: 55e1092a18d90c2a52adf2ce87c8e82826316a81 [formerly e300cebeb86aa237b6f9fc6aaf209f9962c7277c]
Former-commit-id: 93d554dd1566cb8311ed4833015dd9dfac0f37db

+++ Quaver/src/Database/Beatmaps/Beatmap.cs
@@ -121,7 +121,7 @@ namespace Quaver.Database.Beatmaps
         /// <summary>
         ///     The most common bpm for the beatmap
         /// </summary>
-        public decimal Bpm { get; set; }
+        public double Bpm { get; set; }
 
         /// <summary>
         ///     The beatmap's length (Time of the last hit object)
